Refactor: move methods to ui.cpp
