+++ input.cpp
@@ -4,6 +4,7 @@
 #include "config.hpp"
 #include "elona.hpp"
 #include "variables.hpp"
+#include "ui.hpp"
